+++ src/ManagedModels/Filters/FilterPresets.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Thinktomorrow\Chief\ManagedModels\Filters;
 
+use Illuminate\Database\Eloquent\Builder;
 use Thinktomorrow\Chief\ManagedModels\Filters\Presets\InputFilter;
 use Thinktomorrow\Chief\ManagedModels\Filters\Presets\SelectFilter;
 use Thinktomorrow\Chief\ManagedModels\States\PageState\PageState;
@@ -32,6 +33,51 @@ class FilterPresets
         ])->default('');
     }
 
+    /**
+     * Search on a column, relation column or dynamic key.
+     * Relation column can be searched by via relation.column
+     */
+    public static function attribute(string $name, string|array $columns = [], string|array $dynamicKeys = [], string $label = 'Titel', string $dynamicColumn = 'values'): Filter
+    {
+        return InputFilter::make($name, function ($query, $value) use ($dynamicKeys, $columns, $dynamicColumn) {
+            return $query->where(function ($builder) use ($value, $dynamicKeys, $columns, $dynamicColumn) {
+
+                // Extract relation searches
+                foreach ($columns as $i => $column) {
+                    if(false !== strpos($column, '.')) {
+                        [$relation, $columnName] = explode('.', $column);
+
+                        $builder->whereHas($relation, function($query) use($value, $columnName, $dynamicColumn){
+                            return static::queryColumnsOrDynamicAttributes($query->whereRaw('1=0'), $value, [$columnName], [], $dynamicColumn);
+                        });
+
+                        unset($columns[$i]);
+                    }
+                }
+
+                // Columns or dynamic keys
+                return static::queryColumnsOrDynamicAttributes($builder, $value, $columns, $dynamicKeys, $dynamicColumn);
+            });
+        })->label($label);
+    }
+
+    private static function queryColumnsOrDynamicAttributes(Builder $builder, $value, $columns, $dynamicKeys, $dynamicColumn): Builder
+    {
+        foreach ($columns as $column) {
+            $builder->orWhere($column, 'LIKE', '%' . $value . '%');
+        }
+
+        foreach ($dynamicKeys as $dynamicKey) {
+            $dynamicColumnParts = explode('.', $dynamicColumn);
+            $builder->orWhereRaw('LOWER(json_extract(`' . implode('`.`', $dynamicColumnParts) . '`, "$.' . $dynamicKey . '")) LIKE ?', '%' . trim(strtolower($value)) . '%');
+        }
+
+        return $builder;
+    }
+
+    /**
+     * @deprecated use attribute() search instead for an extensive usage range.
+     */
     public static function column(string $name, string|array $columns, ?string $label = null): Filter
     {
         return InputFilter::make($name, function ($query, $value) use ($columns) {
@@ -45,6 +91,9 @@ class FilterPresets
         })->label($label ?? $name);
     }
 
+    /**
+     * @deprecated use attribute() search instead for an extensive usage range.
+     */
     public static function text(string $queryParameter, array $dynamicColumns = ['title'], array $astrotomicColumns = [], string $label = 'Titel', string $jsonColumn = 'values'): Filter
     {
         return InputFilter::make($queryParameter, function ($query, $value) use ($dynamicColumns, $astrotomicColumns, $jsonColumn) {